+++ Model/Resource/Storage/SourceItemStorage.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace BigBridge\ProductImport\Model\Resource\Storage;
+
+/**
+ * Storage for multi-store inventory source items.
+ *
+ * @author Patrick van Bergen
+ */
+class SourceItemStorage
+{
+
+}